Change type of VolumeOfTheTank to double

+++ AutoPark/Vehicles/Vehicle.cs
@@ -21,7 +21,7 @@ namespace AutoPark.Vehicles
         /// <summary>
         /// Volume in liters or kilowatts
         /// </summary>
-        public int VolumeOfTheTank { get; set; }
+        public double VolumeOfTheTank { get; set; }
         public const decimal WeightTaxCoefficient = 0.0013m;
         public decimal TaxPerMonth => Weight * WeightTaxCoefficient + VehicleType.TaxCoefficient * 30 + 5;
